Enemy tracking and abstraction in main is working! Will work on creating a level class that will deal with loading levels with parameters from a file so that the game can be easily changed. Also fixed drawSquare() in the drawing helpers to properly order vertexies for the GL_TRIANGLE_STRIP

+++ src/Enemy.h
@@ -38,7 +38,7 @@ public:
 
 	//update and state setting
 	virtual void update(const double& frameDiff) = 0;
-	virtual void setState(int playerX, int playerY) = 0;
+	virtual void setState(float playerX, float playerY) = 0;
 
 	//drawing
 	virtual void draw() = 0;
